Keep a strong reference to the parent theme
If the `VSThemeLoader` goes out of scope while you have a reference
to a `VSTheme`, then the only reference to the parent theme will be
through the weak `parentTheme` property. This will eventually get
cleaned up, leaving you with an empty reference and no theme
inheritence.

The fix is to make the `parentTheme` property a strong reference.